Fix the spelling of untrimmed
Spelling is hard. It seemed like a good choice at first to replace
vowels that we weren't sure about with apostrophes, but that adds a
bunch of weird symbols to the text. Replace the symbols with actual
letters.

There are five vowels in the English alphabet. "E" seems like a good
choice here.